Add hosted mode
Hosted mode means we are being integrated into some
other image, and in this case, we expect that is ChromeOS.
For now, all it really means is that we create
an initramfs, which does not have any devices, /etc/hosts,
and so on.

Tested by booting and installing non-hosted on chromebooks,
and running testramfs on the hosted initramfs.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>